Change method inputString
